requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
yfinance>=0.2.28
plotly>=5.17.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
